requirements.txt
flask==3.0.3
requests==2.32.3
python-evtx==0.7.4
lxml==5.2.2
urllib3==2.2.2
gunicorn==22.0.0
pytest==8.2.2
